+++ src/layer/riscv/deconvolution_packnto1_fp16s.h
@@ -183,7 +183,7 @@ static void deconvolution_packnto1_fp16sa_rvv(const Mat& bottom_blob, Mat& top_b
 
                 sum = __riscv_vfmv_f_s_f16m1_f16(__riscv_vfredusum_vs_f16m1_f16m1(_sum, __riscv_vfmv_s_f_f16m1(sum, vl), vl));
 
-                sum = activation_ss(sum, activation_type, activation_params);
+                sum = (__fp16)activation_ss(sum, activation_type, activation_params);
 
                 outptr[j] = (__fp16)sum;
             }
